requirements.txt
# Core Django framework
Django>=4.2,<5.0

# WSGI HTTP server for production
gunicorn

# Gestion des fichiers statiques
whitenoise

# Connecteur PostgreSQL
psycopg2-binary

# Gestion de DATABASE_URL
dj-database-url

# Variables d'environnement
python-dotenv
django-environ

# Django REST Framework (si ton projet l'utilise)
djangorestframework

# CORS pour communication avec un frontend
django-cors-headers

# (Optionnel) outils supplémentaires
django-extensions

# Installeation du client mysql
sudo apt install mysql-client -y
#installation de JWT pour les token
# JWT pour les tokens
PyJWT>=2.0.0